+++ src/PhpSpreadsheet/Calculation/Information/Value.php
@@ -52,8 +52,10 @@ class Value
             if (!empty($worksheet) && $cell->getWorksheet()->getParentOrThrow()->getSheetByName($worksheet) === null) {
                 return false;
             }
-            [$column, $row] = Coordinate::indexesFromString($cellValue ?? '');
-            if ($column > 16384 || $row > 1048576) {
+
+            try {
+                [$column, $row] = Coordinate::indexesFromString($cellValue ?? '');
+            } catch (SpreadsheetException) {
                 return false;
             }