+++ src/PhpSpreadsheet/Calculation/LookupRef/RowColumnInformation.php
@@ -5,6 +5,7 @@ namespace PhpOffice\PhpSpreadsheet\Calculation\LookupRef;
 use PhpOffice\PhpSpreadsheet\Calculation\Calculation;
 use PhpOffice\PhpSpreadsheet\Calculation\Information\ErrorValue;
 use PhpOffice\PhpSpreadsheet\Calculation\Information\ExcelError;
+use PhpOffice\PhpSpreadsheet\Cell\AddressRange;
 use PhpOffice\PhpSpreadsheet\Cell\Cell;
 use PhpOffice\PhpSpreadsheet\Cell\Coordinate;
 use PhpOffice\PhpSpreadsheet\Exception as SpreadsheetException;
@@ -131,7 +132,7 @@ class RowColumnInformation
 
     private static function convert0ToName(int|string $result): int|string
     {
-        if (is_int($result) && ($result <= 0 || $result > 1048576)) {
+        if (is_int($result) && ($result <= 0 || $result > AddressRange::MAX_ROW)) {
             return ExcelError::NAME();
         }