+++ src/PhpSpreadsheet/Cell/Coordinate.php
@@ -35,7 +35,11 @@ abstract class Coordinate
     public static function coordinateFromString(string $cellAddress): array
     {
         if (preg_match(self::A1_COORDINATE_REGEX, $cellAddress, $matches)) {
-            return [$matches['col'], $matches['row']];
+            $row = (int) ltrim($matches['row'], '$');
+            // reluctantly allow row 0 due to regression problems
+            if (/*$row > 0 &&*/ $row <= AddressRange::MAX_ROW) {
+                return [$matches['col'], $matches['row']];
+            }
         } elseif (self::coordinateIsRange($cellAddress)) {
             throw new Exception('Cell coordinate string can not be a range of cells');
         } elseif ($cellAddress == '') {
@@ -420,22 +424,28 @@ abstract class Coordinate
                 $indexCache[$columnAddress] = $columnLookup[$columnAddress];
 
                 return $indexCache[$columnAddress];
-            } elseif (!isset($columnAddress[2])) {
+            }
+            if (!isset($columnAddress[2])) {
                 $indexCache[$columnAddress] = $columnLookup[$columnAddress[0]] * 26
                     + $columnLookup[$columnAddress[1]];
 
                 return $indexCache[$columnAddress];
-            } elseif (!isset($columnAddress[3])) {
-                $indexCache[$columnAddress] = $columnLookup[$columnAddress[0]] * 676
+            }
+            if (!isset($columnAddress[3])) {
+                $temp = $columnLookup[$columnAddress[0]] * 676
                     + $columnLookup[$columnAddress[1]] * 26
                     + $columnLookup[$columnAddress[2]];
 
-                return $indexCache[$columnAddress];
+                if ($temp <= AddressRange::MAX_COLUMN_INT) {
+                    $indexCache[$columnAddress] = $temp;
+
+                    return $temp;
+                }
             }
         }
 
         throw new Exception(
-            'Column string index can not be ' . ((isset($columnAddress[0])) ? 'longer than 3 characters' : 'empty')
+            'Column string index can not be ' . ((isset($columnAddress[0])) ? ('beyond ' . AddressRange::MAX_COLUMN) : 'empty')
         );
     }
 
@@ -446,11 +456,19 @@ abstract class Coordinate
      *
      * @param int|numeric-string $columnIndex Column index (A = 1)
      */
-    public static function stringFromColumnIndex(int|string $columnIndex): string
+    public static function stringFromColumnIndex(int|string $columnIndex, bool $tolerateZero = false): string
     {
         /** @var string[] */
         static $indexCache = [];
+        $columnIndex2 = (int) $columnIndex;
+        if ($columnIndex2 === 0 && $tolerateZero) {
+            return '';
+        }
+        if ($columnIndex2 < 1 || $columnIndex2 > AddressRange::MAX_COLUMN_INT) {
+            throw new Exception("Invalid column index $columnIndex");
+        }
 
+        $columnIndex = $columnIndex2;
         if (!isset($indexCache[$columnIndex])) {
             $indexValue = $columnIndex;
             $base26 = '';
@@ -554,7 +572,7 @@ abstract class Coordinate
             $row = 0;
             sscanf($coordinate, '%[A-Z]%d', $column, $row);
             /** @var int $row */
-            $key = (--$row * 16384) + self::columnIndexFromString((string) $column);
+            $key = (--$row * AddressRange::MAX_COLUMN_INT) + self::columnIndexFromString((string) $column);
             $sortKeys[$key] = $coordinate;
         }
         ksort($sortKeys);