+++ src/PhpSpreadsheet/Reader/Xlsx/ColumnAndRowAttributes.php
@@ -2,6 +2,7 @@
 
 namespace PhpOffice\PhpSpreadsheet\Reader\Xlsx;
 
+use PhpOffice\PhpSpreadsheet\Cell\AddressRange;
 use PhpOffice\PhpSpreadsheet\Cell\Coordinate;
 use PhpOffice\PhpSpreadsheet\Reader\DefaultReadFilter;
 use PhpOffice\PhpSpreadsheet\Reader\IReadFilter;
@@ -157,7 +158,7 @@ class ColumnAndRowAttributes extends BaseParserClass
                 for ($columnAddress = $startColumn; $columnAddress !== $endColumn; StringHelper::stringIncrement($columnAddress)) {
                     $columnAttributes[$columnAddress] = $this->readColumnRangeAttributes($column, $readDataOnly);
 
-                    if ((int) ($column['max']) == 16384) {
+                    if ((int) ($column['max']) === AddressRange::MAX_COLUMN_INT) {
                         break;
                     }
                 }