+++ src/PhpSpreadsheet/Style/Style.php
@@ -3,6 +3,7 @@
 namespace PhpOffice\PhpSpreadsheet\Style;
 
 use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Cell\AddressRange;
 use PhpOffice\PhpSpreadsheet\Cell\Coordinate;
 use PhpOffice\PhpSpreadsheet\Chart\ChartColor;
 use PhpOffice\PhpSpreadsheet\Exception;
@@ -131,6 +132,13 @@ class Style extends Supervisor
     {
         return $this->getActiveSheet()->getParentOrThrow();
     }
+
+    private const REGEX_WHOLE_COLUMN = '/^[A-Z]+1:[A-Z]+'
+        . AddressRange::MAX_ROW
+        . '$/';
+    private const REGEX_WHOLE_ROW = '/^A\d+:'
+        . AddressRange::MAX_COLUMN
+        . '\d+$/';
 
     /**
      * Build style array from subcomponents.
@@ -358,12 +366,12 @@ class Style extends Supervisor
 
             // SIMPLE MODE:
             // Selection type, inspect
-            if (preg_match('/^[A-Z]+1:[A-Z]+1048576$/', $pRange)) {
+            if (preg_match(self::REGEX_WHOLE_COLUMN, $pRange)) {
                 $selectionType = 'COLUMN';
 
                 // Enable caching of styles
                 self::$cachedStyles = ['hashByObjId' => [], 'styleByHash' => []];
-            } elseif (preg_match('/^A\d+:XFD\d+$/', $pRange)) {
+            } elseif (preg_match(self::REGEX_WHOLE_ROW, $pRange)) {
                 $selectionType = 'ROW';
 
                 // Enable caching of styles