+++ tests/data/Calculation/LookupRef/ROWonSpreadsheet.php
@@ -17,4 +17,6 @@ return [
     'unknown name as first part of range' => ['#NAME?', 'InvalidCell:A2'],
     'unknown name as second part of range' => ['#NAME?', 'A2:InvalidCell'],
     'qualified name' => [6, 'OtherSheet!localname'],
+    'last valid row' => [1048576, 'A1048576'],
+    'beyond last valid row' => ['#NAME?', 'A1048577'],
 ];